+++ CHANGELOG.md
@@ -292,6 +292,9 @@ Miscellaneous:
   skeleton IDs in a skeleton that is merge into different skeletons by different
   requests.
 
+- Landmark transformations: fix Moving Least Square transformations for skeleton
+  fragments outside of source group bounding box.
+
 - The Connectivity Graph Plot draws now individual bars in each sub-plot
   side-by-side and uses the same neuron names as other widgets (from CATMAID's
   neuron name service).